Made home page "bacteria" button functional
In this build, the bacteria section of home page now leads to a new webpage. This webpage contains details regarding how bacteria works. This includes bacterial structure, how they reproduce, and what harm they may bring to body. This section is important because it helps readers understand possible ways how medicines can be used to combat harmful bacterias.

A limitation of this build is that the other homepage buttons are still not functional.