Made home page "penicillin" button functional
In this build, the bacteria section of home page now leads to a new webpage. This webpage contains details regarding how penicillin works. This includes penicillin structure, how they help combat bacterias, and how bacteria may evolve to resist this drug. This section is important because it helps readers understand properties of an important drug, as well as gives an understanding of bacterial evolution.

A limitation of this build is that not all homepage buttons are functional.