Made home page "macrolide" button functional
In this build, the macrolide section of home page now leads to a new webpage. This webpage contains details regarding how macrolide works. This includes macrolide structure, how they help combat bacterias, and how bacteria may evolve to resist this drug. This section is important because it helps readers understand properties of an important drug, as well as gives an understanding of bacterial evolution.